requirements.txt
aiodns
aiohttp
aiolimiter
asyncio
celery[amqp]
django-celery-beat
datetime
dnspython
duckdb
httpx
ipaddress
kombu
lmdb
loguru
polars
pyarrow
python-dotenv
pyyaml
tldextract
typing
xxhash